LPS-119677 portal-search-elasticsearch7-impl: Merge CompanyIndexCreator logic into CompanyIndexFactory
https://issues.liferay.com/browse/LPS-119677
