+++ modules/apps/portal-search-elasticsearch7/portal-search-elasticsearch7-impl/src/main/java/com/liferay/portal/search/elasticsearch7/internal/ElasticsearchSearchEngine.java
@@ -30,7 +30,6 @@ import com.liferay.portal.kernel.util.Time;
 import com.liferay.portal.kernel.util.Validator;
 import com.liferay.portal.search.ccr.CrossClusterReplicationHelper;
 import com.liferay.portal.search.elasticsearch7.internal.connection.ElasticsearchConnectionManager;
-import com.liferay.portal.search.elasticsearch7.internal.index.CompanyIndexCreator;
 import com.liferay.portal.search.elasticsearch7.internal.index.IndexFactory;
 import com.liferay.portal.search.engine.adapter.SearchEngineAdapter;
 import com.liferay.portal.search.engine.adapter.cluster.ClusterHealthStatus;
@@ -116,7 +115,7 @@ public class ElasticsearchSearchEngine extends BaseSearchEngine {
 
 		_indexFactory.createIndices(restHighLevelClient.indices(), companyId);
 
-		_companyIndexCreator.registerCompanyId(companyId);
+		_indexFactory.registerCompanyId(companyId);
 
 		waitForYellowStatus();
 
@@ -154,7 +153,7 @@ public class ElasticsearchSearchEngine extends BaseSearchEngine {
 			_indexFactory.deleteIndices(
 				restHighLevelClient.indices(), companyId);
 
-			_companyIndexCreator.unregisterCompanyId(companyId);
+			_indexFactory.unregisterCompanyId(companyId);
 		}
 		catch (Exception exception) {
 			if (_log.isWarnEnabled()) {
@@ -206,12 +205,6 @@ public class ElasticsearchSearchEngine extends BaseSearchEngine {
 		super.setIndexWriter(indexWriter);
 	}
 
-	public void unsetCompanyIndexCreator(
-		CompanyIndexCreator companyIndexCreator) {
-
-		_companyIndexCreator = companyIndexCreator;
-	}
-
 	public void unsetCrossClusterReplicationHelper(
 		CrossClusterReplicationHelper crossClusterReplicationHelper) {
 
@@ -260,13 +253,6 @@ public class ElasticsearchSearchEngine extends BaseSearchEngine {
 		}
 
 		return true;
-	}
-
-	@Reference
-	protected void setCompanyIndexCreator(
-		CompanyIndexCreator companyIndexCreator) {
-
-		_companyIndexCreator = companyIndexCreator;
 	}
 
 	@Reference(
@@ -371,7 +357,6 @@ public class ElasticsearchSearchEngine extends BaseSearchEngine {
 	private static final Log _log = LogFactoryUtil.getLog(
 		ElasticsearchSearchEngine.class);
 
-	private CompanyIndexCreator _companyIndexCreator;
 	private CrossClusterReplicationHelper _crossClusterReplicationHelper;
 	private ElasticsearchConnectionManager _elasticsearchConnectionManager;
 	private IndexFactory _indexFactory;
