+++ modules/apps/portal-search-elasticsearch7/portal-search-elasticsearch7-impl/src/main/java/com/liferay/portal/search/elasticsearch7/internal/index/IndexFactory.java
@@ -25,4 +25,8 @@ public interface IndexFactory {
 
 	public void deleteIndices(IndicesClient indicesClient, long companyId);
 
+	public void registerCompanyId(long companyId);
+
+	public void unregisterCompanyId(long companyId);
+
 }
